openai server trait and move HTTP endpoints to crates/runtime/src/http/v1/
