+++ crates/llms/src/openai/mod.rs
@@ -14,7 +14,11 @@ limitations under the License.
 use crate::embeddings::{Embed, EmbeddingInput, Error as EmbedError, Result as EmbedResult};
 use crate::nql::{Error as NqlError, Nql, Result as NqlResult};
 
-use async_openai::types::{CreateEmbeddingRequest, CreateEmbeddingRequestArgs};
+use async_openai::error::OpenAIError;
+use async_openai::types::{
+    CreateChatCompletionRequest, CreateChatCompletionResponse, CreateEmbeddingRequest,
+    CreateEmbeddingRequestArgs, CreateEmbeddingResponse,
+};
 use async_openai::{
     config::OpenAIConfig,
     types::{
@@ -27,7 +31,10 @@ use async_openai::{
 use async_trait::async_trait;
 use futures::future::try_join_all;
 use serde_json::Value;
+use server::Server;
 use snafu::ResultExt;
+
+pub mod server;
 
 const MAX_COMPLETION_TOKENS: u16 = 1024_u16; // Avoid accidentally using infinite tokens. Should think about this more.
 
@@ -214,5 +221,22 @@ fn to_openai_embedding_input(input: EmbeddingInput) -> OpenAiEmbeddingInput {
         EmbeddingInput::Tokens(t) => OpenAiEmbeddingInput::IntegerArray(t),
         EmbeddingInput::StringBatch(sb) => OpenAiEmbeddingInput::StringArray(sb),
         EmbeddingInput::TokensBatch(tb) => OpenAiEmbeddingInput::ArrayOfIntegerArray(tb),
+    }
+}
+
+#[async_trait]
+impl Server for Openai {
+    async fn chat(
+        &mut self,
+        req: CreateChatCompletionRequest,
+    ) -> Result<CreateChatCompletionResponse, OpenAIError> {
+        self.client.chat().create(req).await
+    }
+
+    async fn embed(
+        &mut self,
+        req: CreateEmbeddingRequest,
+    ) -> Result<CreateEmbeddingResponse, OpenAIError> {
+        self.client.embeddings().create(req).await
     }
 }
