+++ crates/runtime/src/http/v1/oai.rs
@@ -1,0 +1,43 @@
+/*
+Copyright 2024 The Spice.ai OSS Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+     https://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+use std::sync::Arc;
+
+use axum::{
+    body::Bytes,
+    extract::Path,
+    http::StatusCode,
+    response::{IntoResponse, Response},
+    Extension,
+};
+
+use crate::datafusion::DataFusion;
+
+pub(crate) async fn chat(
+    Path(name): Path<String>,
+    Extension(df): Extension<Arc<DataFusion>>,
+    body: Bytes,
+) -> Response {
+    (StatusCode::INTERNAL_SERVER_ERROR, "not implemented").into_response()
+}
+
+pub(crate) async fn embeddings(
+    Path(name): Path<String>,
+    Extension(df): Extension<Arc<DataFusion>>,
+    body: Bytes,
+) -> Response {
+    (StatusCode::INTERNAL_SERVER_ERROR, "not implemented").into_response()
+}
